refactor: update naming
